+++ src/DistanceMatrix/johnson.hpp
@@ -1,6 +1,7 @@
 #include "constants.h"
 #include <boost/config.hpp>
 #include <boost/graph/adjacency_list.hpp>
+#include <cstddef>
 
 namespace johnson {
 
@@ -13,14 +14,14 @@ typedef std::pair<unsigned int, unsigned int> Edge;
 
 typedef struct graph {
   unsigned int V;
-  unsigned int E;
+  size_t E;
   Edge *edge_array;
   dist_t *weights;
 } graph_t;
 
 graph_t *johnson_init(std::vector<dist_t>);
 
-graph_t *johnson_init2(const unsigned int n, const double p, const unsigned long seed);
+graph_t *johnson_init2(const size_t n, const double p, const unsigned long seed);
 
 typedef struct edge {
   unsigned int u;
